I…have no memory of why I added `masto-accounts-token` here.
It looks…clearly old, seeing as it's not using `http` (and that'd
easily take care of a lot of the stuff it does manually).

I'll leave it, for now (in case I find it somewhere in the Masto
documentation), but there're definitely other endpoints I should get
added that I've neglected.

I'm not sure what the token will look like, for
`masto-account-create`, (a simple string would make most sense, to me,
but I can't guarantee that) so I'll need to test this, at some point.

We'll just leave it returning the thing directly, for now.